pyproject.toml
[tool.poetry]
name = "raspberry-pi-shop"
version = "0.1.0"
description = ""
authors = ["Atanas Pozharliev <ABPozharliev19@codingburgas.bg>"]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
Django = "^4.1.2"
psycopg2 = "^2.9.4"
django-cron = "^0.6.0"
Scrapy = "^2.7.0"
scrapy-djangoitem = "^1.1.1"
django-cors-headers = "^3.13.0"
djangorestframework-camel-case = "^1.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"